refactor: rename folders and files
