+++ src/components/ui/form/searchField/SearchField.tsx
@@ -6,8 +6,8 @@ import { variantStyles } from '@/constants/borderVariants.constants'
 import { X } from 'lucide-react'
 import { LoaderCircle } from 'lucide-react'
 import { useDebounce } from '@/hooks/useDebounce'
-import type { SearchResult } from '@/services/transactionApi/transaction.types'
-import { useAssetSearch } from '@/services/transactionApi/useAssetSearch'
+import type { SearchResult } from '@/services/ISINsearchApi/ISINsearch.types'
+import { useAssetSearch } from '@/services/ISINsearchApi/useAssetSearch'
 import { useAsset } from '@/store/useAsset'
 import type { ITransactionForm } from '@/components/ModalTransactions/TransactionForm/transactionForm.types'
