+++ src/services/ISINsearchApi/apiSearchISIN.ts
@@ -6,14 +6,14 @@ import type {
     AssetIdentifier,
     AssetPrice,
     TransactionDate,
-} from './transaction.types'
+} from './ISINsearch.types'
 import {
     transformSearchResults,
     transformCurrentPriceSearchResults,
     transformHistoricalPriceSearchResults,
     transformDataForUrl,
-} from './transaction.utils'
-import { CRYPTO_PROVIDERS, BOND_PROVIDERS, STOCK_ETF_PROVIDERS, API_CONFIGS, API_KEYS } from './transaction.config'
+} from './ISINsearch.utils'
+import { CRYPTO_PROVIDERS, BOND_PROVIDERS, STOCK_ETF_PROVIDERS, API_CONFIGS, API_KEYS } from './ISINsearch.config'
 
 // Search function with fallback
 export const searchAssets = async (query: string, assetType: AssetType): Promise<SearchResult[]> => {
